+++ ragger-tests/test_sign_cmd.py
@@ -214,6 +214,61 @@ def test_sign_tx_unknown_tx_rejected(backend, scenario_navigator, firmware, navi
 
     assert len(e.value.data) == 0
 
+# can blind sign a transfer transaction with too many inputs
+def test_sign_tx_blind_sign_big_transfer_tx(backend, scenario_navigator, firmware, navigator):
+    client = Client(backend, use_block_protocol=True)
+    path = "m/44'/784'/0'"
+
+    _, public_key, _, _ = client.get_public_key(path=path)
+
+    transaction = base64.b64decode('AAAAAAA2ACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dACASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXQAgElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV0AIBJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dAQEBAAEAABJV0JMExcyCS570/ZGW9bjpYWe0+wJm9w3TaCmtUZ1dAsC7lrrw/2YMlYTpcvws/Y2vBrIdxoKXuS8umgDP8CjtBxVYHgAAAAAgwsfZ/KrW4rPd+o3wBKj6xX93MXmD82nDn57olpXDLGca710L3H+IcWY/kHTo0fu6GdLrfUQkAi1wVrqP3n9MGORpzBwAAAAAIJKL0Cjd/7Uh4bM8xMiutEO/R5Out2S5tOP10PW9yGwZElXQkwTFzIJLnvT9kZb1uOlhZ7T7Amb3DdNoKa1RnV3uAgAAAAAAAMD2JQAAAAAAAA==')
+
+    object_list = [base64.b64decode('AAEB5GnMHAAAAAAoGu9dC9x/iHFmP5B06NH7uhnS631EJAItcFa6j95/TBic4wYAAAAAAAASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXSCMWQXAJLoXR3VqNIzOUarpAW+dd5yJz4WVe+HOMs0HBGATDwAAAAAA')
+      , base64.b64decode('AAEBBxVYHgAAAAAowLuWuvD/ZgyVhOly/Cz9ja8Gsh3Ggpe5Ly6aAM/wKO0i+gMNAAAAAAASVdCTBMXMgkue9P2RlvW46WFntPsCZvcN02gprVGdXSCz/wDyVt9e0DZ6dGsI7AOxGIB/ghCPBinKM6ZMFqQPLmATDwAAAAAA')
+                   ]
+
+    def apdu_task():
+        return client.sign_tx(path=path, transaction=transaction, object_list=object_list)
+
+    def nav_task():
+        if firmware.device.startswith("nano"):
+            navigator.navigate_and_compare(
+                instructions=[ NavInsID.RIGHT_CLICK # Warning...
+                               , NavInsID.RIGHT_CLICK, NavInsID.RIGHT_CLICK # Transaction Hash
+                               , NavInsID.BOTH_CLICK]
+                , timeout=10
+                , path=scenario_navigator.screenshot_path
+                , test_case_name=scenario_navigator.test_name
+                , screen_change_before_first_instruction=False
+                , screen_change_after_last_instruction=False
+            )
+        else:
+            # Dismiss the "Blind signing ahead" screen
+            navigator.navigate(
+                instructions=[NavInsID.USE_CASE_CHOICE_REJECT]
+                , timeout=20
+                , screen_change_before_first_instruction=True
+                , screen_change_after_last_instruction=True
+            )
+            # Below is similar to scenario_navigator.review_approve()
+            # But screen_change_before_first_instruction=True causes hang
+            navigator.navigate_until_text_and_compare(
+                navigate_instruction=NavInsID.SWIPE_CENTER_TO_LEFT
+                , validation_instructions=[NavInsID.USE_CASE_REVIEW_CONFIRM, NavInsID.USE_CASE_STATUS_DISMISS]
+                , text="^Hold to sign$"
+                , timeout=20
+                , path=scenario_navigator.screenshot_path
+                , test_case_name=scenario_navigator.test_name
+                , screen_change_before_first_instruction=False
+                , screen_change_after_last_instruction=True
+            )
+
+    def check_result(result):
+        assert len(result) == 64
+        assert check_signature_validity(public_key, result, transaction)
+
+    with blind_sign_enabled(firmware, navigator):
+        run_apdu_and_nav_tasks_concurrently(apdu_task, nav_task, check_result)
 @contextmanager
 def blind_sign_enabled(firmware, navigator):
     toggle_blind_sign(firmware, navigator)