+++ rust-app/src/parser/common.rs
@@ -57,7 +57,7 @@ pub const IOTA_SYSTEM_STATE_ID: CoinID = [
 // This does not contain the correct module and function names, as we don't have a way to create const ArrayVec with them
 pub const IOTA_COIN_TYPE: CoinType = (IOTA_COIN_ID, ArrayVec::new_const(), ArrayVec::new_const());
 
-pub const IOTA_COIN_DIVISOR: u8 = 9;
+pub const IOTA_COIN_DECIMALS: u8 = 9;
 
 pub type ObjectData = CoinData;